fix: accept other filenames for LICENSE
The added filenames are LICENCE and COPYING, both of which are
commonly used. There might be others accepted by GitHub, but the
list can easily be extended.

This check could be a little slower due to having to do three times
the number of string comparisons, but hoefully that is counteracted
by the switch to using a generator, ensuring that an intermediate
list doesn't have to be built, and that the check can terminate as
soon as a file is found.

Fixes https://github.com/scientific-python/cookie/issues/326