Remove `ref-names` from `.git_archival.txt` (#444)
It turns out that the `ref-names` field in `.git_archival.txt` will generate a different value depending on whether the commit an archive was built from was the head of a branch at the time of archiving. That's a problem because it means someone trying to compare the hash signatures of two archives built from the same commit might get different results based on *when* the archives were made, and not on any other intrinsic feature of the archives' content. This removes the field in order to make Git archives a bit more stable and useful.